+++ demos/ca_demos_web/web/cellular_automata.dart
@@ -12,43 +12,30 @@ import 'package:cellular_automata/rules.dart';
 import 'package:stagexl/src/ui/color.dart';
 
 // Fully featured example of using cellular_automata
-void startSimulation(
-    {CARules rules,
-    num worldWidth,
-    num worldHeight,
-    num stageWidth,
-    num stageHeight,
-    num speedMs,
-    MathematicalGenerators mathematicalGenerator,
-    StageXLDisplayMode displayMode,
-    CanvasElement canvas}) {
+void startSimulation({
+  CARules rules,
+  num worldWidth,
+  num worldHeight,
+  num stageWidth,
+  num stageHeight,
+  num speedMs,
+  StageXLDisplayMode displayMode,
+  CanvasElement canvas,
+  CellWorld world,
+  Map palette,
+  CAGenerator generator,
+}) {
   print('Cellular Automata Demo');
   print('World: ${worldWidth}x$worldHeight');
   print('Stage: ${stageWidth}x$stageHeight');
   print('Speed: $speedMs');
-
-  final world = new CellWorld<GameOfLifeStates>(
-      width: worldWidth,
-      height: worldHeight,
-      defaultState: GameOfLifeStates.DEAD);
-
-  final palette = new Map<GameOfLifeStates, int>.from({
-    GameOfLifeStates.DEAD: Color.Blue,
-    GameOfLifeStates.DEAD_UNDER_POPULATED: Color.DarkBlue,
-    GameOfLifeStates.DEAD_OVER_POPULATED: Color.BlueViolet,
-    GameOfLifeStates.ALIVE: Color.Yellow,
-    GameOfLifeStates.ALIVE_BORN: Color.LightYellow,
-  });
 
   final sim = new Simulator(
       world: world,
       rules: rules,
       generationDuration: new Duration(milliseconds: speedMs),
       palette: palette,
-      generator: new MathematicalGenerator<GameOfLifeStates>(
-          type: mathematicalGenerator,
-          valueTrue: GameOfLifeStates.ALIVE_BORN,
-          valueFalse: GameOfLifeStates.DEAD));
+      generator: generator);
 
   final renderer = new StageXLRenderer(width: worldWidth, height: worldHeight)
     ..initStageXL(
@@ -78,17 +65,13 @@ void _initSimulation([dynamic _]) {
   num stageHeight;
 
   CARules rules;
-  switch (params['rules']) {
-    case 'game_of_life':
-    default:
-      rules = new GameOfLife();
-
-      break;
-  }
+  CellWorld world;
+  Map palette;
+  CAGenerator generator;
 
   final num renderSize = int.parse(params['render_size'] ?? '8');
 
-  MathematicalGenerators generator;
+  MathematicalGenerators generatorType;
   if (params['generator'] != null) {
     final gen = params['generator'].toString().toUpperCase();
 
@@ -98,7 +81,7 @@ void _initSimulation([dynamic _]) {
         gens.add(E.toString().substring(E.toString().indexOf('\.') + 1)));
 
     if (gens.contains(gen))
-      generator = MathematicalGenerators.values[gens.indexOf(gen)];
+      generatorType = MathematicalGenerators.values[gens.indexOf(gen)];
   }
 
   StageXLDisplayMode displayMode;
@@ -124,13 +107,71 @@ void _initSimulation([dynamic _]) {
 
   querySelector('body').classes.add(bodyClass);
 
+  switch (params['rules' ?? 'game_of_life']) {
+    case 'game_of_life':
+      rules = new GameOfLife();
+      world = new CellWorld<GameOfLifeStates>(
+          width: worldWidth,
+          height: worldHeight,
+          defaultState: GameOfLifeStates.DEAD);
+
+      palette = new Map<GameOfLifeStates, int>.from({
+        GameOfLifeStates.DEAD: Color.Blue,
+        GameOfLifeStates.DEAD_UNDER_POPULATED: Color.DarkBlue,
+        GameOfLifeStates.DEAD_OVER_POPULATED: Color.BlueViolet,
+        GameOfLifeStates.ALIVE: Color.Yellow,
+        GameOfLifeStates.ALIVE_BORN: Color.LightYellow,
+      });
+
+      generator = new MathematicalGenerator<GameOfLifeStates>(
+          type: generatorType,
+          valueTrue: GameOfLifeStates.ALIVE_BORN,
+          valueFalse: GameOfLifeStates.DEAD);
+      break;
+
+    case 'game_of_life_simple':
+      rules = new GameOfLifeSimple();
+      world = new CellWorld<bool>(
+          width: worldWidth, height: worldHeight, defaultState: false);
+
+      palette = new Map<bool, int>.from({
+        false: Color.DarkRed,
+        true: Color.GreenYellow,
+      });
+
+      generator = new MathematicalGenerator<bool>(
+          type: generatorType, valueTrue: true, valueFalse: false);
+      break;
+
+    case 'brians_brain':
+      rules = new BriansBrain();
+      world = new CellWorld<BriansBrainStates>(
+          width: worldWidth,
+          height: worldHeight,
+          defaultState: BriansBrainStates.OFF);
+
+      palette = new Map<BriansBrainStates, int>.from({
+        BriansBrainStates.OFF: Color.DarkOliveGreen,
+        BriansBrainStates.DYING: Color.LawnGreen,
+        BriansBrainStates.ON: Color.OrangeRed,
+      });
+
+      generator = new MathematicalGenerator<BriansBrainStates>(
+          type: generatorType,
+          valueTrue: BriansBrainStates.ON,
+          valueFalse: BriansBrainStates.OFF);
+      break;
+  }
+
   startSimulation(
       rules: rules,
       worldWidth: worldWidth,
       worldHeight: worldHeight,
       stageHeight: stageHeight,
       stageWidth: stageWidth,
-      mathematicalGenerator: generator,
+      generator: generator,
+      world: world,
+      palette: palette,
       displayMode: displayMode,
       speedMs: speedMs,
       canvas: querySelector('#stage')