+++ lib/rules.dart
@@ -3,3 +3,5 @@ library cellular_automata.rules;
 
 export 'package:cellular_automata/src/rules/_ca_rules.dart';
 export 'package:cellular_automata/src/rules/game_of_life.dart';
+export 'package:cellular_automata/src/rules/game_of_life_simple.dart';
+export 'package:cellular_automata/src/rules/brians_brain.dart';