+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.3
+
+- Added 'Conway's Game of Life Simple' Rules - has binary states and only 24 lines of code
+- Added 'Brian's Brain' Rules
+- Updated demos
+
+
 ## 0.2
 
 - Optimised cell processing, now only processes cells that neighbor active states